+++ lib/src/response/autosuggest.dart
@@ -1,4 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:what3words/src/response/coordinates.dart';
 
 import 'response.dart';
 part 'autosuggest.g.dart';
@@ -37,4 +38,42 @@ class Suggestion {
       _$SuggestionFromJson(json);
 
   Map<String, dynamic> toJson() => _$SuggestionToJson(this);
+}
+
+///Encapulates AutoSuggest suggestions
+@JsonSerializable(explicitToJson: true)
+class AutosuggestWithCoordinates extends Response<AutosuggestWithCoordinates> {
+  List<SuggestionWithCoordinates> suggestions;
+
+  AutosuggestWithCoordinates({this.suggestions});
+
+  factory AutosuggestWithCoordinates.fromJson(Map<String, dynamic> json) =>
+      _$AutosuggestWithCoordinatesFromJson(json);
+
+  Map<String, dynamic> toJson() => _$AutosuggestWithCoordinatesToJson(this);
+}
+
+@JsonSerializable(explicitToJson: true)
+class SuggestionWithCoordinates {
+  String country;
+  String nearestPlace;
+  String words;
+  CoordinatesResponse coordinates;
+  int distanceToFocusKm;
+  int rank;
+  String language;
+
+  SuggestionWithCoordinates(
+      {this.country,
+        this.nearestPlace,
+        this.words,
+        this.coordinates,
+        this.distanceToFocusKm,
+        this.rank,
+        this.language});
+
+  factory SuggestionWithCoordinates.fromJson(Map<String, dynamic> json) =>
+      _$SuggestionWithCoordinatesFromJson(json);
+
+  Map<String, dynamic> toJson() => _$SuggestionWithCoordinatesToJson(this);
 }