+++ lib/src/service/what3words_v3.dart
@@ -1,6 +1,9 @@
+import 'package:what3words/src/request/autosuggest_options.dart';
+import 'package:what3words/src/request/autosuggest_selection_request.dart';
+import 'package:what3words/src/request/autosuggest_with_coordinates_request.dart';
+
 import '../request/autosuggest_request.dart';
 import '../request/available_languages_request.dart';
-import '../request/bounding_box.dart';
 import '../request/convert_to_3wa_request.dart';
 import '../request/convert_to_coordinates_request.dart';
 import '../request/coordinate.dart';
@@ -72,13 +75,16 @@ class What3WordsV3 {
   ///Returns a section of the 3m x 3m what3words grid for a bounding box. The bounding box is specified by lat,lng,lat,lng
   ///as south,west,north,east.
   ///
-  ///[boundingBox] [BoundingBox], for which the grid should be returned. The requested box must not exceed 4km
+  ///[southwest], SouthWest point of the bounding box for which the grid should be returned. The requested box must not exceed 4km
+  ///from corner to corner. Latitudes must be &gt;= -90 and &lt;= 90, but longitudes are allowed to wrap around 180. To specify a
+  ///bounding-box that crosses the anti-meridian, use longitude greater than 180.
+  ///[northeast], NorthEast point of the bounding box for which the grid should be returned. The requested box must not exceed 4km
   ///from corner to corner. Latitudes must be &gt;= -90 and &lt;= 90, but longitudes are allowed to wrap around 180. To specify a
   ///bounding-box that crosses the anti-meridian, use longitude greater than 180.
   ///
   ///returns a [GridSectionRequestBuilder] instance suitable for invoking a `grid-section` API request
-  GridSectionRequestBuilder gridSection(BoundingBox boundingBox) {
-    return GridSectionRequestBuilder(this, boundingBox);
+  GridSectionRequestBuilder gridSection(Coordinates southwest, Coordinates northeast) {
+    return GridSectionRequestBuilder(this, southwest, northeast);
   }
 
   ///AutoSuggest can take a slightly incorrect 3 word address, and suggest a list of valid 3 word addresses. It has powerful
@@ -87,10 +93,37 @@ class What3WordsV3 {
   ///
   ///[input] The full or partial 3 word address to obtain suggestions for. At minimum this must be the first two complete
   ///words plus at least one character from the third word.
+  ///[options] The autosuggest options and clippings, check available options here https://docs.what3words.com/api/v3/#autosuggest
   ///
   ///returns a [AutosuggestRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestRequestBuilder autosuggest(String input) {
-    return AutosuggestRequestBuilder(this, input);
+  AutosuggestRequestBuilder autosuggest(String input, {AutosuggestOptions options}) {
+    return AutosuggestRequestBuilder(this, input, options);
+  }
+
+  ///AutoSuggestWithCoordinates can take a slightly incorrect 3 word address, and suggest a list of valid 3 word addresses including coordinates. It has powerful
+  ///features which can, for example, optionally limit results to a country or area, and prefer results which are near the user.
+  ///For full details, please see the complete API documentation at https://docs.what3words.com/api/v3/#autosuggest
+  ///
+  ///[input] The full or partial 3 word address to obtain suggestions for. At minimum this must be the first two complete
+  ///words plus at least one character from the third word.
+  ///[options] The autosuggest options, check available options here https://docs.what3words.com/api/v3/#autosuggest
+  ///
+  ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
+  AutosuggestWithCoordinatesRequestBuilder autosuggestWithCoordinates(String input, {AutosuggestOptions options}) {
+    return AutosuggestWithCoordinatesRequestBuilder(this, input, options);
+  }
+
+  ///autosuggest-selection enables simple reporting for what3words address selections in accounts.what3words.com.
+  ///It should be called once in conjunction with autosuggest or autosuggest-with-coordinates.
+  ///when an end user picks a what3words address from the suggestions presented to them.
+  ///[rawInput] The full or partial 3 word address used on the autosuggest or autosuggest-with-coordinates call.
+  ///[sourceApi] The source of the selected autosuggest, can be 'text' or 'voice'.
+  ///[words] The 3 word address of the selected suggestion.
+  ///[rank] The rank of the selected suggestion.
+  ///[options] The autosuggest options used on the autosuggest or autosuggest-with-coordinates call.
+  ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
+  AutosuggestSelectionRequestBuilder autosuggestSelection(String rawInput, String sourceApi, String words, int rank, {AutosuggestOptions options}) {
+    return AutosuggestSelectionRequestBuilder(this, rawInput, sourceApi, words, rank, options);
   }
 
   What3WordsV3Service what3words() {